_hooksの型定義の修正
