+++ src/app/_hooks/useCategoryBudgetSummary.ts
@@ -1,14 +1,5 @@
 import useFetch from "@/app/_hooks/useFetch";
-
-export interface CategoryBudgetSummary {
-  categoryId: number;
-  name: string;
-  budgetAmount: number;
-  displayColor: string;
-  totalExpense: number;
-  percent: number;
-  isOver: boolean;
-}
+import type { CategoryBudgetSummary } from "@/app/_types/CategoryBudgetSummary";
 
 const useCategoryBudgetSummary = (year: number, month: number) => useFetch<CategoryBudgetSummary[]>(`/api/category_budgets/summary?year=${year}&month=${month}`);
