+++ src/app/_hooks/useExpenses.ts
@@ -1,20 +1,5 @@
 import useFetch from "@/app/_hooks/useFetch";
-
-export interface Expense {
-  id: number;
-  categoryId: number;
-  amount: number;
-  date: string;
-  note?: string;
-  createdAt: string;
-  updatedAt: string;
-  category: {
-    id: number;
-    name: string;
-    budgetAmount: number;
-    displayColor: string;
-  };
-}
+import type { Expense } from "@/app/_types/Expense";
 
 const useExpenses = () => useFetch<{ expenses: Expense[] }>("/api/expenses");
