+++ src/app/_types/CategoryBudget.ts
@@ -1,0 +1,8 @@
+export interface CategoryBudget {
+  id: number;
+  name: string;
+  budgetAmount: number;
+  displayColor: string;
+  createdAt: string;
+  updatedAt: string;
+}
