+++ src/app/_types/CategoryBudgetSummary.ts
@@ -1,0 +1,9 @@
+export interface CategoryBudgetSummary {
+  categoryId: number;
+  name: string;
+  budgetAmount: number;
+  displayColor: string;
+  totalExpense: number;
+  percent: number;
+  isOver: boolean;
+}
